Update README.md
In this repository, there will be R codes and Powerpoint slides for the students of Plant Pathology and Microbiology who want to learn basic R programming. We will start with the simple arithmetic calculations, and then slowly we will go for advanced studies. We will make histograms, boxplots, figures, charts, and more focusing on the topics related to plant pathology and microbiology.